refactor(MovieReservation): model 패키징 분리
