+++ app/src/main/java/woowacourse/movie/presentation/reservation/booking/MovieReservationMapper.kt
@@ -4,6 +4,9 @@ import woowacourse.movie.model.MovieReservation
 import woowacourse.movie.model.date.ScreenDateTime
 import woowacourse.movie.model.date.ScreenDateTimes
 import woowacourse.movie.model.date.ScreeningMovie
+import woowacourse.movie.presentation.reservation.booking.model.MovieReservationUiState
+import woowacourse.movie.presentation.reservation.booking.model.ScreenDateTimeUiModel
+import woowacourse.movie.presentation.reservation.booking.model.ScreeningMovieUiModel
 import woowacourse.movie.presentation.reservation.result.ReservationResultUiModel
 import java.time.LocalDate
 import java.time.LocalDateTime
