+++ app/src/main/java/woowacourse/movie/presentation/reservation/booking/model/ScreenDateTimeUiModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.presentation.reservation.booking
+package woowacourse.movie.presentation.reservation.booking.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
